Move fastRestore function from fdbserver into fdbbackup
fdbbackup cannot depend on fdbserver.
So we cannot use functions defined in fdbserver in fdbbackup or fdbclient.
We move the fastRestore function that initialize the restore request from fdbserver to fdbclient.

It is compilable now.
