+++ fdbserver/Restore.actor.cpp
@@ -4266,137 +4266,4 @@ ACTOR Future<std::string> RestoreConfig::getProgress_impl(Reference<RestoreConfi
 }
 
 
-////-------Restore Agent: Kick off the restore by sending the restore requests
-ACTOR static Future<ERestoreState> waitFastRestore(Database cx, Key tagName, bool verbose) {
-	// MX: We should wait on all restore before proceeds
-	printf("Wait for restore to finish\n");
-	state int waitNum = 0;
-	state ReadYourWritesTransaction tr2(cx);
-	state Future<Void> watch4RestoreRequestDone;
-	loop {
-		try {
-			tr2.reset();
-			tr2.setOption(FDBTransactionOptions::ACCESS_SYSTEM_KEYS);
-			tr2.setOption(FDBTransactionOptions::LOCK_AWARE);
 
-			watch4RestoreRequestDone = tr2.watch(restoreRequestDoneKey);
-			wait( tr2.commit() );
-			printf("[INFO] Finish setting up watch for restoreRequestDoneKey\n");
-			break;
-		} catch( Error &e ) {
-			TraceEvent("CheckRestoreRequestDoneErrorMX").detail("ErrorInfo", e.what());
-			printf("[WARNING] Transaction error: setting up watch for restoreRequestDoneKey, error:%s\n", e.what());
-			wait( tr2.onError(e) );
-		}
-	}
-
-	loop {
-		try {
-			tr2.reset();
-			tr2.setOption(FDBTransactionOptions::ACCESS_SYSTEM_KEYS);
-			tr2.setOption(FDBTransactionOptions::LOCK_AWARE);
-			Optional<Value> restoreRequestDoneKeyValue = wait( tr2.get(restoreRequestDoneKey) );
-			if ( restoreRequestDoneKeyValue.present() ) {
-				printf("!!! restoreRequestTriggerKey has been set before we wait on the key: Restore has been done before restore agent waits for the done key\n");
-				break;
-			}
-			wait(watch4RestoreRequestDone);
-			printf("[INFO] watch for restoreRequestDoneKey is triggered\n");
-			break;
-		} catch( Error &e ) {
-			TraceEvent("CheckRestoreRequestDoneErrorMX").detail("ErrorInfo", e.what());
-			//printf("[WARNING]  Transaction error: waiting for the watch of the restoreRequestDoneKey, error:%s\n", e.what());
-			wait( tr2.onError(e) );
-		}
-	}
-
-	loop {
-		try {
-			tr2.reset();
-			tr2.setOption(FDBTransactionOptions::ACCESS_SYSTEM_KEYS);
-			tr2.setOption(FDBTransactionOptions::LOCK_AWARE);
-			state Optional<Value> numFinished = wait(tr2.get(restoreRequestDoneKey));
-			if (numFinished.present()) {
-				int num = decodeRestoreRequestDoneValue(numFinished.get());
-				TraceEvent("RestoreRequestKeyDoneFinished").detail("NumFinished", num);
-				printf("[INFO] RestoreRequestKeyDone, numFinished:%d\n", num);
-			}
-			printf("[INFO] RestoreRequestKeyDone: clear the key in a transaction");
-			tr2.clear(restoreRequestDoneKey);
-			// NOTE: The clear transaction may fail in uncertain state. We need to retry to clear the key
-			wait( tr2.commit() );
-			break;
-		} catch( Error &e ) {
-			TraceEvent("CheckRestoreRequestDoneErrorMX").detail("ErrorInfo", e.what());
-			printf("[WARNING] Clearing the restoreRequestDoneKey has error in transaction: %s. We will retry to clear the key\n", e.what());
-			wait( tr2.onError(e) );
-		}
-
-	}
-
-	printf("MX: Restore is finished\n");
-
-	return ERestoreState::COMPLETED;
-
-}
-
-
-ACTOR static Future<Version> _fastRestore(Database cx, Key tagName, Key url, bool waitForComplete, Version targetVersion, bool verbose, KeyRange range, Key addPrefix, Key removePrefix) {
-	state Reference<IBackupContainer> bc = IBackupContainer::openContainer(url.toString());
-	state BackupDescription desc = wait(bc->describeBackup());
-	wait(desc.resolveVersionTimes(cx));
-
-	printf("Backup Description\n%s", desc.toString().c_str());
-	if(targetVersion == invalidVersion && desc.maxRestorableVersion.present())
-		targetVersion = desc.maxRestorableVersion.get();
-
-	Optional<RestorableFileSet> restoreSet = wait(bc->getRestoreSet(targetVersion));
-
-	if(!restoreSet.present()) {
-		TraceEvent(SevWarn, "FileBackupAgentRestoreNotPossible")
-			.detail("BackupContainer", bc->getURL())
-			.detail("TargetVersion", targetVersion);
-		fprintf(stderr, "ERROR: Restore version %lld is not possible from %s\n", targetVersion, bc->getURL().c_str());
-		throw restore_invalid_version();
-	}
-
-	if (verbose) {
-		printf("Restoring backup to version: %lld\n", (long long) targetVersion);
-	}
-
-	// NOTE: The restore agent makes sure we only support 1 restore range for each restore request for now!
-	// The simulation test did test restoring multiple restore ranges in one restore request though.
-	state Reference<ReadYourWritesTransaction> tr(new ReadYourWritesTransaction(cx));
-	state int restoreIndex = 0;
-	loop {
-		try {
-			tr->setOption(FDBTransactionOptions::ACCESS_SYSTEM_KEYS);
-			tr->setOption(FDBTransactionOptions::LOCK_AWARE);
-			Standalone<StringRef> restoreTag(tagName.toString() + "_" + std::to_string(restoreIndex));
-			bool locked = true;
-			struct RestoreRequest restoreRequest(restoreIndex, restoreTag, KeyRef(bc->getURL()), true, targetVersion, true, range, Key(), Key(), locked, g_random->randomUniqueID());
-			tr->set(restoreRequestKeyFor(restoreRequest.index), restoreRequestValue(restoreRequest));
-			tr->set(restoreRequestTriggerKey, restoreRequestTriggerValue(1)); //backupRanges.size = 1 because we only support restoring 1 range in real mode
-			wait(tr->commit()); //Trigger MX restore
-			break;
-		} catch(Error &e) {
-			if(e.code() != error_code_restore_duplicate_tag) {
-				wait(tr->onError(e));
-			}
-		}
-	}
-
-	if(waitForComplete) {
-		ERestoreState finalState = wait(waitFastRestore(cx, tagName, verbose));
-		if(finalState != ERestoreState::COMPLETED)
-			throw restore_error();
-	}
-
-	return targetVersion;
-}
-
-
-ACTOR Future<Version> fastRestore(Database cx, Standalone<StringRef>  tagName, Standalone<StringRef> url, bool waitForComplete, long targetVersion, bool verbose, Standalone<KeyRangeRef> range, Standalone<StringRef> addPrefix, Standalone<StringRef> removePrefix) {
-	Version targetVersion =  wait( _fastRestore(cx, tagName, url, waitForComplete, targetVersion, verbose, range, addPrefix, removePrefix) );
-	return targetVersion;
-}
